add endpoint get and patch user

+++ models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, Integer
+from sqlalchemy import String, Integer, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.ext.declarative import declarative_base
 
@@ -8,7 +8,7 @@ class User(Base):
     __tablename__ = "users"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
-    telegram_id: Mapped[int] = mapped_column(Integer, index=True, unique=True)
+    telegram_id: Mapped[int] = mapped_column(BigInteger, index=True, unique=True)
     twitch_nickname: Mapped[str] = mapped_column(String, unique=True, index=True)
     first_name: Mapped[str] = mapped_column(String)
     telegram_username: Mapped[str] = mapped_column(String, nullable=True)
